Fix code review issues: stale status metadata, dedup types, extract helpers
- Fix critical bug: optimistic status update now clears enriched metadata
  (statusGroup/statusLabel/statusColor) so getStatusDisplay() recomputes
  from DEFAULT_STATUS_CONFIG instead of showing stale values
- Extract DEFAULT_STATUS_CONFIG and TaskStatusGroup to @/lib/task-status-config
  to fix cross-layer import (API route was importing from component path)
- Deduplicate TaskStatusGroup: single source of truth in shared lib,
  re-exported by task-list-types.ts and types.ts
- Extract getStatusDisplay() and getAssigneeText() into task-helpers.ts
  to eliminate duplicated status/assignee display logic across components
- Add clarifying comment for .name vs .label property mismatch in API
  enrichment (DB configs use .name, DEFAULT_STATUS_CONFIG uses .label)
