+++ apps/web/src/components/layout/middle-content/page-views/task-list/task-list-types.ts
@@ -1,3 +1,7 @@
+// Re-export from shared lib (single source of truth for status config)
+import { DEFAULT_STATUS_CONFIG, type TaskStatusGroup } from '@/lib/task-status-config';
+export { DEFAULT_STATUS_CONFIG, type TaskStatusGroup } from '@/lib/task-status-config';
+
 // Shared types and config for task list views
 
 export interface TaskStatusConfig {
@@ -81,16 +85,7 @@ export interface TaskListData {
 }
 
 export type TaskPriority = 'low' | 'medium' | 'high';
-export type TaskStatusGroup = 'todo' | 'in_progress' | 'done';
 export type ViewMode = 'table' | 'kanban';
-
-// Default status configs used when a task list has no custom configs yet
-export const DEFAULT_STATUS_CONFIG: Record<string, { label: string; color: string; group: TaskStatusGroup }> = {
-  pending: { label: 'To Do', color: 'bg-slate-100 text-slate-700 dark:bg-slate-800 dark:text-slate-300', group: 'todo' },
-  in_progress: { label: 'In Progress', color: 'bg-amber-100 text-amber-700 dark:bg-amber-900 dark:text-amber-300', group: 'in_progress' },
-  completed: { label: 'Done', color: 'bg-green-100 text-green-700 dark:bg-green-900 dark:text-green-300', group: 'done' },
-  blocked: { label: 'Blocked', color: 'bg-red-100 text-red-700 dark:bg-red-900 dark:text-red-300', group: 'in_progress' },
-};
 
 // Build a STATUS_CONFIG-compatible map from custom status configs
 export function buildStatusConfig(configs: TaskStatusConfig[]): Record<string, { label: string; color: string; group: TaskStatusGroup }> {
