+++ apps/web/src/components/tasks/TaskDetailSheet.tsx
@@ -29,12 +29,13 @@ import { cn } from '@/lib/utils';
 import { MultiAssigneeSelect } from '@/components/layout/middle-content/page-views/task-list/MultiAssigneeSelect';
 import { DueDatePicker } from '@/components/layout/middle-content/page-views/task-list/DueDatePicker';
 import {
-  DEFAULT_STATUS_CONFIG,
   PRIORITY_CONFIG,
   STATUS_ORDER,
   type TaskPriority,
 } from '@/components/layout/middle-content/page-views/task-list/task-list-types';
+import { DEFAULT_STATUS_CONFIG } from '@/lib/task-status-config';
 import type { Task } from './types';
+import { getStatusDisplay } from './task-helpers';
 
 export interface TaskDetailSheetProps {
   task: Task | null;
@@ -75,12 +76,10 @@ export function TaskDetailSheet({
 
   if (!task) return null;
 
-  const isCompleted = task.statusGroup ? task.statusGroup === 'done' : task.status === 'completed';
+  const statusDisplay = getStatusDisplay(task);
+  const isCompleted = statusDisplay.group === 'done';
   const hasLinkedPage = Boolean(task.pageId && task.driveId);
-
-  // Safe status display
-  const statusLabel = task.statusLabel || DEFAULT_STATUS_CONFIG[task.status]?.label || task.status;
-  const statusColor = task.statusColor || DEFAULT_STATUS_CONFIG[task.status]?.color || 'bg-slate-100 text-slate-600';
+  const { label: statusLabel, color: statusColor } = statusDisplay;
 
   const startEditTitle = () => {
     setEditingTitle(task.title);
